fix(left-manage-validator): separate into different folders
