+++ src/pages/ManageValidator/LeftManage/LeftManage.tsx
@@ -4,8 +4,8 @@ import { InformationBox } from '@status-im/components'
 import { CloseCircleIcon } from '@status-im/icons'
 
 import Header from './Header'
-import ValidatorDataCards from './ValidatorDataCards'
-import ValidatorSettingsCards from './ValidatorSettingsCards'
+import ValidatorDataCards from './ValidatorDataCards/ValidatorDataCards'
+import ValidatorSettingsCards from './ValidatorSettingsCards/ValidatorSettingsCards'
 import ValidatorGraffiti from './ValidatorGraffiti'
 import ValidatorDataTabs from './ManageValidatorTable/ValidatorDataTabs'
 import Footer from './Footer'
