+++ src/pages/ManageValidator/LeftManage/ValidatorSettingsCards/ValidatorSettingCard.tsx
@@ -3,7 +3,7 @@ import { DisconnectIcon, AdvancedIcon, ExternalIcon } from '@status-im/icons'
 import { useNavigate } from 'react-router-dom'
 import { Stack, YStack } from 'tamagui'
 
-import { CARD_TITLES } from '../../../constants'
+import { CARD_TITLES } from '../../../../constants'
 
 type CardTitle = (typeof CARD_TITLES)[number]
